Broke C code into separate files, moved common stuff to php_aws_crt.h (#25)
* Broke C code into separate files, moved common stuff to php_aws_crt.h

* clang-format

* Use a unity build to make PHP do the right thing with our lib
