+++ ext/awscrt.c
@@ -1,0 +1,13 @@
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
+/* This is a unity-build style source file, as PHP's build system is simplest with 1 source file per extension */
+
+#include "credentials.c"
+#include "crt.c"
+#include "event_loop.c"
+#include "http.c"
+#include "signing.c"
+#include "stream.c"
